requirements.txt
asgiref==3.10.0
crispy-bootstrap5==2025.6
Django==5.2.7
django-crispy-forms==2.4
python-decouple==3.8
python-http-client==3.3.7
sendgrid==3.6.5
sendgrid-django==4.2.0
sqlparse==0.5.3
whitenoise==6.11.0
